refactor: API 호출 함수 내 fetcher 적용 및 tmdb 내 통합
