+++ types/movie.types.ts
@@ -1,5 +1,3 @@
-import { fetcher } from '@/lib/fetcher';
-
 export interface TMDBResponse {
 	page: number;
 	results: Movie[];
@@ -18,14 +16,4 @@ export interface Movie {
 	poster_path: string;
 	vote_average: number;
 	vote_count: number;
-}
-
-export async function getMoviesApi(endpoint: string): Promise<TMDBResponse | null> {
-	try {
-		const movieData = await fetcher(endpoint);
-		return movieData;
-	} catch (error) {
-		console.error('Error fetching movie data:', error);
-		return null;
-	}
 }
